Moving tests to take advantage of DBPath (ensuring cleanup)...
