+++ src/compaction_filter.rs
@@ -101,34 +101,3 @@ where
         }
     }
 }
-
-#[cfg(test)]
-#[allow(unused_variables)]
-fn test_filter(level: u32, key: &[u8], value: &[u8]) -> Decision {
-    use self::Decision::*;
-    match key.first() {
-        Some(&b'_') => Remove,
-        Some(&b'%') => Change(b"secret"),
-        _ => Keep,
-    }
-}
-
-#[test]
-fn compaction_filter_test() {
-    use {Options, DB};
-
-    let path = "_rust_rocksdb_filtertest";
-    let mut opts = Options::default();
-    opts.create_if_missing(true);
-    opts.set_compaction_filter("test", test_filter);
-    {
-        let db = DB::open(&opts, path).unwrap();
-        let _ = db.put(b"k1", b"a");
-        let _ = db.put(b"_k", b"b");
-        let _ = db.put(b"%k", b"c");
-        db.compact_range(None, None);
-        assert_eq!(&*db.get(b"k1").unwrap().unwrap(), b"a");
-        assert!(db.get(b"_k").unwrap().is_none());
-        assert_eq!(&*db.get(b"%k").unwrap().unwrap(), b"secret");
-    }
-}
